+++ components/my-page/MyOneGameResult.tsx
